refactoring so that modal now is in CreateManifestModal, and former CreateManifestFormContainer content is located there as well
