+++ src/components/CreateManifestForm/index.ts
@@ -1,12 +1,10 @@
 import CreateManifestForm from './CreateManifestForm';
-import CreateManifestFormContainer from './CreateManifestFormContainer';
 import CreateManifestFormError from './CreateManifestFormError';
 import CreateManifestFormLoading from './CreateManifestFormLoading';
 import CreateManifestFormSuccess from './CreateManifestFormSuccess';
 
 export {
   CreateManifestForm,
-  CreateManifestFormContainer,
   CreateManifestFormError,
   CreateManifestFormLoading,
   CreateManifestFormSuccess
