+++ src/components/CreateManifestModal/__tests__/CreateManifestModal.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CreateManifestFormContainer from '../CreateManifestFormContainer';
+import CreateManifestModal from '../CreateManifestModal';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { render } from '@testing-library/react';
 import useSatelliteVersions, { SatelliteVersion } from '../../../hooks/useSatelliteVersions';
@@ -9,17 +9,24 @@ jest.mock('../../../hooks/useSatelliteVersions');
 const queryClient = new QueryClient();
 
 describe('Create Manifest Form Container', () => {
-  it('renders correctly once satellite versions have been loaded', () => {
+  it('renders correctly once satellite versions have been loaded', async () => {
     (useSatelliteVersions as jest.Mock).mockReturnValue({
       body: [] as SatelliteVersion[],
-      isError: false
+      isError: false,
+      isLoading: false
     });
-    const props = { handleModalToggle: (): any => null };
+    const props = { handleModalToggle: (): any => null, isModalOpen: true };
     const { container } = render(
       <QueryClientProvider client={queryClient}>
-        <CreateManifestFormContainer {...props} />
+        <CreateManifestModal {...props} />
       </QueryClientProvider>
     );
-    expect(container).toMatchSnapshot();
+
+    /**
+     * document.body needed because the modal
+     * does not render within the container
+     *
+     */
+    expect(document.body).toMatchSnapshot();
   });
 });
