+++ src/components/CreateManifestModal/index.ts
@@ -1,0 +1,3 @@
+import CreateManifestModal from './CreateManifestModal';
+
+export { CreateManifestModal as default };
